Put lids on the barrels
- Added the ability to evaluate, modify and block data operations immediately before data is written to or read from a barrel's backing data store.
- Modified existing, and added new, unit tests to take changes into account. All tests passing

To use the new feature, create a class that implements the new ILid interface and assign it to your datastore-specific Barrel class's new Lid property. The ILid interface requires implementations of the AddingToBarrel and GettingFromBarrel methods.  See the MonkeyCache.TestsShared.TestLid class for an example.

Return the string that was passed to the methods if no modifications are needed, or modify the received data as desired before returning the changes. Returning null from either method causes the associated Barrel operation to be aborted.